Implement Ad-hoc analyzers
Creating Ad-hoc analyzers on the fly is now supported.  A tokenizer and one
or more filters may be specified, which will then query ES for the analysis.
Ordering of the filters are maintained since this affects the final output.

Unfortunately, this commit includes some substantial overhead in the form
of jQuery and Select2.js.  However, the functionality of the 'tagging' Select2
is perfect for the Filter selection...so I'll just live with the extra
overhead.

Closes #6